+++ src/Validation/dtos/validation.isdate.dto.ts
@@ -1,0 +1,16 @@
+import { ValidationGeneralDto } from "./validation.general.dto";
+
+/**
+ * Defaults to {
+ *     format:"YYYY-MM-DD",
+*      delimiters:["-"],
+ *     strictMode:false
+ * }
+ */
+export interface IValidationIsDateDto extends Omit<ValidationGeneralDto, 'params' | 'customFunction'> {
+    params?: {
+        format?: string,
+        delimiters?: string[],
+        strictMode?: boolean
+    }
+}